// studio-ui/ui/app/src/models/ContentType.ts
export interface ContentTypeFieldValidation<T = any> {
	id: ValidationKeys | DescriptorFieldValidationKeys;
	value: T;
	level: 'required' | 'suggestion';
}

export type ValidationKeys =
	| 'allowedContentTypeTags'
	| 'allowedContentTypes'
	| 'allowedEmbeddedContentTypes'
	| 'allowedSharedContentTypes'
	| 'allowedSharedExistingContentTypes'
	| 'minCount'
	| 'maxCount'
	| 'maxLength'
	| 'readOnly'
	| 'required'
	| 'width'
	| 'height'
	| 'minWidth'
	| 'minHeight'
	| 'maxWidth'
	| 'maxHeight'
	| 'minValue'
	| 'maxValue'
	| 'dropTargetsNotFound'
	| 'registerNotFound'
	| 'allowImagesFromRepo'
	| 'allowImageUpload'
	| 'allowVideosFromRepo'
	| 'allowVideoUpload'
	| 'allowAudioUpload'
	| 'allowAudioFromRepo'
	| 'pattern'
	| 'allowDuplicates'
	| 'addMedia'
	| 'minSize';

export type ContentTypeFieldValidations = Record<ValidationKeys, ContentTypeFieldValidation>;

export interface ValidationResult {
	id?: string;
	message?: string;
	level?: 'required' | 'suggestion' | 'info';
	values?: object;
}

export interface ContentTypeField {
	id: string;
	name: string;
	description?: string;
	type: string;
	sortable?: boolean;
	validations: Partial<ContentTypeFieldValidations>;
	properties?: Partial<
		// region
		{ plugin: LegacyFormDefinitionField['plugin'] } & Record<
			// These are just some of the possibilities...
			| 'size'
			| 'maxlength'
			| 'readonly'
			| 'allowEditWithoutWarning'
			| 'tokenized'
			| 'escapeContent'
			| 'minSize'
			| 'maxSize'
			| 'itemManager'
			| 'disableFlattening'
			| 'useSingleValueFilename'
			| 'contentTypes'
			| 'tags'
			| 'useMVS'
			| 'rows'
			| 'allowResize'
			| 'width'
			| 'height'
			| 'thumbnailWidth'
			| 'thumbnailHeight'
			| 'imageManager'
			| 'forceRootBlockPTag'
			| 'forcePTags'
			| 'forceBRTags'
			| 'supportedChannels'
			| 'rteConfiguration'
			| 'videoManager'
			| 'datasource'
			| 'selectAll'
			| 'listDirection'
			| 'showDate'
			| 'showTime'
			| 'showClear'
			| 'showNowLink'
			| 'populate'
			| 'allowPastDate'
			| 'populateDateExp'
			| 'useCustomTimezone'
			| 'readonlyEdit'
			| 'cols'
			| 'minOccurs'
			| 'maxOccurs'
			| 'emptyvalue'
			| 'maxValue'
			| 'minValue'
			| 'autoGrow'
			| 'fileManager'
			| string,
			{
				name: string;
				value: string | boolean | number;
				type: string;
			}
		>
		// endregion
	>;
	defaultValue: any;
	fields?: LookupTable<ContentTypeField>;
	values?: { label: string; value: string }[];
	helpText?: string;
	// localized: boolean;
	// disabled: boolean;
	// readOnly: boolean;
	// mapsTo: string;
	// mapsToType: string;
	// mapsToTarget: string;
	// editor: string;
}

export interface ContentTypeRepeatField extends ContentTypeField {
	fields: LookupTable<ContentTypeField>;
}

export interface NewContentTypeField extends ContentTypeField {
	NEW?: true;
}

export interface FormDefinitionPlugin {
	type: string;
	name: string;
	filename: string;
	pluginId: string;
}

export interface DataSource {
	id: string;
	type: string;
	title: string;
	interface: string;
	properties: LookupTable;
	plugin?: FormDefinitionPlugin;
}

export interface NewDataSource extends DataSource {
	NEW?: true;
}

export type LegacyComponentType = 'component' | 'page' | 'file';

export interface ContentType {
	id: string;
	name: string;
	description: string;
	type: LegacyComponentType;
	quickCreate: boolean;
	quickCreatePath: string;
	displayTemplate: string;
	mergeStrategy: string;
	// ∨∨∨ Added during TypeBuilder 2 ∨∨∨
	// icon
	hasJsController: boolean; // TODO: Should this one be the file name, like `thumbnailFileName`?
	thumbnailFileName: string;
	isHeadless: boolean; // TODO: Consider removing this property.
	// TODO: Renaming this property to something more meaningful. "Paths" of what?? e.g. allowedCreationPaths?
	//  Possibly split into `allowedTypeCreationBlacklist` and `allowedTypeCreationWhitelist`?
	//  `creationPathBlacklist`, `creationPathWhitelist`?
	paths: {
		excludes?: { pattern: string[] };
		includes?: { pattern: string[] };
	};
	'delete-dependencies': {
		'delete-dependency': {
			pattern: string;
			'remove-empty-folder': boolean;
		}[];
	};
	'copy-dependencies': {
		'copy-dependency': {
			pattern: string;
			target: string;
		}[];
	};
	previewable: boolean;
	// ^^^ Added during TypeBuilder 2 ^^^
	sections: ContentTypeSection[];
	fields: LookupTable<ContentTypeField>;
	dataSources: DataSource[];
}
// ...
export interface ContentTypeSection {
	// ∨∨∨ Added during TypeBuilder 2 ∨∨∨
	id: string;
	color: string;
	// ^^^ Added during TypeBuilder 2 ^^^
	title: string;
	fields: string[];
	description: string;
	expandByDefault: boolean;
}

export interface LegacyFormDefinitionProperty {
	label?: string; // => display name
	name: string; // => id
	type: string;
	value: string;
}

export interface LegacyFormDefinitionField {
	id: string;
	title: string;
	description: string;
	defaultValue: string;
	type: string;
	help: string;
	iceId: string;
	constraints: { constraint: LegacyFormDefinitionProperty | Array<LegacyFormDefinitionProperty> };
	properties: { property: LegacyFormDefinitionProperty | Array<LegacyFormDefinitionProperty> };
	fields?: { field: LegacyFormDefinitionField | Array<LegacyFormDefinitionField> };
	plugin: FormDefinitionPlugin;
	// Repeat groups carry these both at the top and inside "properties" (duplicated)
	minOccurs?: string;
	maxOccurs?: string;
}